Port Vivado to new Edalize API (#303)
* Port the vivado flow to use the new Edalize API
