+++ edalize/tools/templates/vivado/vivado-synth.tcl.j2
@@ -1,0 +1,3 @@
+launch_runs synth_1{{ jobs }}
+wait_on_run synth_1
+exit [regexp -nocase -- {synth_design (error|failed)} [get_property STATUS [get_runs synth_1]] match]
