+++ tests/test_vivado/minimal/Makefile
@@ -1,8 +1,10 @@
 #Auto generated by Edalize
 
-all: test_vivado_minimal_0.bit
+all: post_build
 
-test_vivado_minimal_0.xpr: test_vivado_minimal_0.tcl
+pre_build: 
+
+test_vivado_minimal_0.xpr: test_vivado_minimal_0.tcl | pre_build
 	$(EDALIZE_LAUNCHER) vivado -notrace -mode batch -source test_vivado_minimal_0.tcl
 
 test_vivado_minimal_0.runs/synth_1/__synthesis_is_complete__: test_vivado_minimal_0_synth.tcl test_vivado_minimal_0.xpr
@@ -18,3 +20,11 @@ build-gui: test_vivado_minimal_0.xpr
 
 pgm: test_vivado_minimal_0_pgm.tcl test_vivado_minimal_0.bit
 	$(EDALIZE_LAUNCHER) vivado -quiet -nolog -notrace -mode batch -source test_vivado_minimal_0_pgm.tcl -tclargs xc7a35tcsg324-1 test_vivado_minimal_0.bit
+
+post_build: test_vivado_minimal_0.bit
+
+pre_run: 
+
+run: pre_run
+
+post_run: run
