+++ tests/test_vivado/yosys/Makefile
@@ -1,12 +1,14 @@
 #Auto generated by Edalize
 
-all: test_vivado_yosys_0.bit
+all: post_build
+
+pre_build: 
+
+test_vivado_yosys_0.edif: edalize_yosys_template.tcl | pre_build
+	$(EDALIZE_LAUNCHER) yosys -l yosys.log -p 'tcl edalize_yosys_template.tcl'
 
 test_vivado_yosys_0.xpr: test_vivado_yosys_0.tcl test_vivado_yosys_0.edif
 	$(EDALIZE_LAUNCHER) vivado -notrace -mode batch -source test_vivado_yosys_0.tcl
-
-test_vivado_yosys_0.edif: edalize_yosys_template.tcl
-	$(EDALIZE_LAUNCHER) yosys -l yosys.log -p 'tcl edalize_yosys_template.tcl'
 
 synth: test_vivado_yosys_0.edif
 
@@ -18,3 +20,11 @@ build-gui: test_vivado_yosys_0.xpr
 
 pgm: test_vivado_yosys_0_pgm.tcl test_vivado_yosys_0.bit
 	$(EDALIZE_LAUNCHER) vivado -quiet -nolog -notrace -mode batch -source test_vivado_yosys_0_pgm.tcl -tclargs xc7a35tcsg324-1 test_vivado_yosys_0.bit
+
+post_build: test_vivado_yosys_0.bit
+
+pre_run: 
+
+run: pre_run
+
+post_run: run
