+++ tests/test_vivado/yosys/edalize_yosys_procs.tcl
@@ -1,3 +1,6 @@
+proc load_plugins {} {
+}
+
 proc read_files {} {
 
 }
