+++ tests/test_vivado/yosys/edalize_yosys_template.tcl
@@ -1,5 +1,7 @@
 yosys -import
 source edalize_yosys_procs.tcl
+load_plugins
+yosys -import
 
 verilog_defaults -push
 verilog_defaults -add -defer
